refactor: split rag retrieval and orchestration
